Add ctor.http to SharedClass
The SharedClass has the `http` property always set now. If the
ctor does not provide `http.path`, a default value of `'/' + name` is
used. This way we can decouple the REST (HTTP) routing from the model
name and make the REST path accessible from any external code.

The second part of the commit modifies ext/swagger to use this new
property.