Exclude body-parser from the browser bundle
When the express dependency was upgraded from 3.x to 4.x,
adapters were changed to require `body-parser` directly, instead
of using `express.json` and `express.urlencoded`.

The module `body-parser` does not work in browser, as it is loading
a list of mime types from a file.

This commit adds a package.json rule to exlude `body-parser` from
the browser bundle.