Add `options.rest.xml`, disable XML by default
Add an option `rest.xml` that quickly enables/disables support
for responses in XML format.

The XML support is disabled by default now.

Versions `2.7.1` and newer has XML output enabled by default,
which causes requests made from a browser to be handled with
XML response type.

Such behaviour is considered as a breaking change when compared
to `2.7.0`. This commit restores the behaviour of `2.7.0` again.