Fixate jayson version to prevent regression
The recently released version jayson@1.2.1 introduced a breaking change
in https://github.com/tedeh/jayson/commit/dc0a13dc4 which is causing
our json-rpc adapter to fail, because we are passing a zero-argument
wrapper function to jayson.

This commit is fixating jayson version to 1.2.0, which is the latest
version working for us.