Add remote invocation phases
Rework "invokeMethodInContext" to use a concept of phases that allows
external components a more fine-grained way for extending built-in
functionality.

The new API is exposed on `remotes.phases`, which is an instance
of PhaseList from loopback-phase.

There are two built-in phases available: "auth" and "invoke".

Hooks registered via `remotes.authorization`, `remotes.before` and
`remotes.after` are invoked in "auth", "invoke:before" and
"invoke:after" phases.

Here is the complete order in which different hooks/handler are called:

 - hook "authorization"
 - phase handlers "auth"
 - hook "remotes.before"
 - phase handlers "invoke:before"
 - invoke method
 - phase handlers "invoke:use"
 - hook "remotes.after"
 - phase handlers "invoke:after"