Reject array/date values for object arguments
Having to distinguish between an array and a plain object in remote
methods is tedious. For example, LoopBack Model constructors expect
an object as the first argument, but will accept an array too, with
undefined result.

This commit changes the implementation of object converter to reject
values that are of a different strong-remoting type,
namely "array" and "date"