Refuse to override builtin file type
Arguments of { type: 'file' } are handled specially by strong-remoting
and don't go through the usual TypeConverter path.

This causes problems in LoopBack applications with a model called
"File", which registers a type converter for "file" (the name is case
insensitive).

In this commit, we are modifying TypeRegistry to throw an error when
the caller attempts to override the built-in "file" type.