fix: destroy stream result on client close
Close the stream result returned by a remote method when the HTTP
client closes the HTTP connection. This is important for long-living
or infinite streams, for example LoopBack's change stream, in order
to allow them to release resources.

Because `ReadableStream#destroy()` was added only in Node.js 8.0.0,
we detect whether the `destroy()` method is available before
enabling this fix.