Add support for pass-through authentication
Allow the rest connector to call other loopback services by passing
on the user authorization from the options where available.
This is the fallback if no other authorization is specified.

This allows loopback servers to communicate securely with other loopback
servers by passing on the auth token. Useful for using loopback as
microservices.

This new options is disabled by default, it can be enabled by adding
the following "options" to the datasource configuration:

    {
      "connector": "remote",
      // ...
      "options": {
        "passAccessToken": "true"
      }
    }