Close event streams when requested as JSON
Before this change, when the client requested a change event stream
as a newline-delimited JSON stream, the original event stream was not
closed when the client closed the connection. As a result, LoopBack
applications were leaking memory.

With this change in place, the server event stream is correctly closed
after the client terminated the connection.