Account for / routes in RestAdapter.allRoutes
Previously, there was a bug in RestAdapter.allRoutes where the paths for
routes which were mounted at / would return as empty instead of the
expected / because of some trailing slash removal logic that was faulty.
This commit fixes this so that it accounts for cases where the path
itself is /, so the logic will leave those unchanged.

Fixes strongloop/strong-remoting#487